requirements.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
aiohttp==3.6.3
astor==0.8.1
async-timeout==3.0.1
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1625831763871/work
chardet==3.0.4
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.3
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
coverage @ file:///C:/ci/coverage_1614614921360/work
cryptography @ file:///C:/ci/cryptography_1616769514516/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython @ file:///C:/ci/cython_1626257069493/work
dataclasses==0.8
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
discord.py==1.5.0
gast==0.2.2
google-auth @ file:///opt/conda/conda-bld/google-auth_1646735974934/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///C:/ci/grpcio_1614884461394/work
h5py==2.10.0
idna==2.8
idna-ssl @ file:///C:/ci/idna_ssl_1611752657819/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1631916885413/work
ipython @ file:///C:/ci/ipython_1593446240034/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi==0.17.0
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver @ file:///C:/ci/kiwisolver_1612282446297/work
Markdown @ file:///C:/ci/markdown_1614364003005/work
matplotlib @ file:///C:/ci/matplotlib-suite_1613408055530/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==4.7.6
numpy @ file:///C:/ci/numpy_and_numpy_base_1603480701039/work
oauthlib @ file:///opt/conda/conda-bld/oauthlib_1644332107998/work
olefile==0.46
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==6.2.1
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.17.2
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
PyJWT @ file:///C:/ci/pyjwt_1619651879817/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyreadline==2.1
PySocks @ file:///C:/ci/pysocks_1605305839978/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==0.20.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scipy==1.3.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==1.14.0
tensorboard-plugin-wit==1.6.0
tensorflow==1.15.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
torch==1.7.1+cu110
tornado @ file:///C:/ci/tornado_1606942379977/work
traitlets @ file:///C:/ci/traitlets_1632759765830/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3==1.25.11
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==0.16.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306197271/work
wincertstore==0.2
wrapt==1.12.1
yarl==1.5.1
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
